Set stderr to empty string when using terminal on Windows.
    Windows HCSShim requires that stderr is an empty string when using terminal.
	Reference: https://github.com/microsoft/hcsshim/blob/200feabd854da69f615a598ed6a1263ce9531676/cmd/containerd-shim-runhcs-v1/service_internal.go#L127
